Update package declaration
