+++ package/src/lib.rs
@@ -23,5 +23,6 @@ pub use errors::*;
 pub mod imports;
 pub mod inputs;
 pub mod outputs;
+pub mod package;
 pub mod root;
 pub mod source;
