+++ package/src/package.rs
@@ -1,0 +1,27 @@
+use crate::errors::PackageError;
+
+use serde::Deserialize;
+use std::path::PathBuf;
+
+#[derive(Deserialize)]
+pub struct Package {
+    pub name: String,
+    pub version: String,
+    pub description: Option<String>,
+    pub license: Option<String>,
+}
+
+impl Package {
+    pub fn new(package_name: &str) -> Self {
+        Self {
+            name: package_name.to_owned(),
+            version: "0.1.0".to_owned(),
+            description: None,
+            license: None,
+        }
+    }
+
+    pub fn remove_package(_package_name: &str) -> Result<(), PackageError> {
+        Ok(())
+    }
+}
